The last major syscall separation commit completely broke our lseek() as well as the linux emulated lseek().  It's sheer luck that the system works at all :-).  Fix lseek's 64 bit return value.

+++ sys/sys/kern_syscall.h
@@ -25,7 +25,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $DragonFly: src/sys/sys/kern_syscall.h,v 1.10 2003/11/03 15:57:34 daver Exp $
+ * $DragonFly: src/sys/sys/kern_syscall.h,v 1.11 2003/11/10 20:57:17 dillon Exp $
  */
 
 #ifndef _SYS_KERN_SYSCALL_H_
@@ -119,7 +119,7 @@ int kern_ftruncate(int fd, off_t length);
 int kern_futimes(int fd, struct timeval *tptr);
 int kern_getdirentries(int fd, char *buf, u_int count, long *basep, int *res);
 int kern_link(struct nameidata *nd, struct nameidata *linknd);
-int kern_lseek(int fd, off_t offset, int whence, int *res);
+int kern_lseek(int fd, off_t offset, int whence, off_t *res);
 int kern_mkdir(struct nameidata *nd, int mode);
 int kern_mknod(struct nameidata *nd, int mode, int dev);
 int kern_open(struct nameidata *nd, int flags, int mode, int *res);
